containers/publisher/requirements.txt
requests
google-api-python-client
tweepy
configparser
google-cloud-pubsub
python-dateutil
pylint
